backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
groq==0.8.0
python-dotenv==1.0.1
httpx==0.27.2 